# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flashcards API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/flashcards/{flashcard}/tags/{tag}": {
            "post": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Attach a tag to a flashcard",
                "description": "Attach a tag",
                "operationId": "6ccef4256e554333866e9caeec0d8e92",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "flashcard"
                ],
                "summary": "Detach a tag",
                "description": "Detach a tag from a flashcard",
                "operationId": "4aa1cfd0e7806637ef415f2125731791",
                "parameters": [
                    {
                        "name": "flashcard",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "403": {
                        "description": "Not permitted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/tags": {
            "get": {
                "tags": [
                    "tag"
                ],
                "summary": "List tags",
                "operationId": "4d0f4baf9e6fa992c0d85fbf140ebf17",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tag"
                ],
                "summary": "Create a tag",
                "operationId": "a0a246f29dcb119a2189fd230fc10028",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "colour"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mathematics"
                                    },
                                    "colour": {
                                        "type": "string",
                                        "example": "green"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tags/{tag}": {
            "put": {
                "tags": [
                    "tag"
                ],
                "summary": "Update a tag",
                "operationId": "41b49eb4f2a830e150f50252f4d831b8",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "colour"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mathematics"
                                    },
                                    "colour": {
                                        "type": "string",
                                        "example": "green"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Tag not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tags/{id}": {
            "delete": {
                "tags": [
                    "tag"
                ],
                "summary": "Delete a tag",
                "operationId": "6f092682e8bbdea016bb928c862fc3e4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "operationId": "2f25885f3b6f61ca11724ec302d2e439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "password_confirmation",
                                    "display_name"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "password"
                                    },
                                    "display_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/user/{user}": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Get logged-in user details",
                "operationId": "a0fa3a08f20fb611866a52d1bb2e43dd",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}